chore: refactor code to be more readable
